// frontend/src/apisv2beta1/artifact/models/V2beta1IOType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * backend/api/v2beta1/artifact.proto
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: version not set
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * Describes the I/O relationship between
 * Artifacts/Parameters and Tasks.
 * There are a couple of instances where
 * input/outputs have special types such
 * as in the case of LoopArguments or
 * dsl.Collected outputs.
 *
 *  - UNSPECIFIED: For validation
 *  - COMPONENT_DEFAULT_INPUT: This is used for inputs that are
 * provided via default parameters in
 * the component input definitions
 *  - TASK_OUTPUT_INPUT: This is used for inputs that are
 * provided via upstream tasks.
 * In the sdk this appears as:
 * TaskInputsSpec.kind.task_output_parameter
 * & TaskInputsSpec.kind.task_output_artifact
 *  - COMPONENT_INPUT: Used for inputs that are
 * passed from parent tasks.
 *  - RUNTIME_VALUE_INPUT: Hardcoded values passed
 * as arguments to the task.
 *  - COLLECTED_INPUTS: Used for dsl.Collected
 * Usage of this type indicates that all
 * Artifacts within the IOArtifact.artifacts
 * are inputs collected from sub tasks with
 * ITERATOR_OUTPUT outputs.
 *  - ITERATOR_INPUT: In a for loop task, introduced via ParallelFor, this type
 * is used to indicate whether this resolved input belongs
 * to a parameterIterator or artifactIterator.
 * In such a case the "artifacts" field for IOArtifact.artifacts
 * is the list of resolved items for this parallelFor.
 *  - ITERATOR_INPUT_RAW: Hardcoded iterator parameters.
 * Raw Iterator inputs have no producer
 *  - ITERATOR_OUTPUT: When an output is produced by a Runtime Iteration Task
 * This value is use to differentiate between standard inputs
 *  - OUTPUT: All other output types fall under this type.
 *  - ONE_OF_OUTPUT: An output of a Conditions branch.
 * @export
 */
export const V2beta1IOType = {
  UNSPECIFIED: 'UNSPECIFIED',
  COMPONENT_DEFAULT_INPUT: 'COMPONENT_DEFAULT_INPUT',
  TASK_OUTPUT_INPUT: 'TASK_OUTPUT_INPUT',
  COMPONENT_INPUT: 'COMPONENT_INPUT',
  RUNTIME_VALUE_INPUT: 'RUNTIME_VALUE_INPUT',
  COLLECTED_INPUTS: 'COLLECTED_INPUTS',
  ITERATOR_INPUT: 'ITERATOR_INPUT',
  ITERATOR_INPUT_RAW: 'ITERATOR_INPUT_RAW',
  ITERATOR_OUTPUT: 'ITERATOR_OUTPUT',
  OUTPUT: 'OUTPUT',
  ONE_OF_OUTPUT: 'ONE_OF_OUTPUT',
  TASK_FINAL_STATUS_OUTPUT: 'TASK_FINAL_STATUS_OUTPUT',
} as const;
export type V2beta1IOType = (typeof V2beta1IOType)[keyof typeof V2beta1IOType];

export function instanceOfV2beta1IOType(value: any): boolean {
  for (const key in V2beta1IOType) {
    if (Object.prototype.hasOwnProperty.call(V2beta1IOType, key)) {
      if (V2beta1IOType[key as keyof typeof V2beta1IOType] === value) {
        return true;
      }
    }
  }
  return false;
}

export function V2beta1IOTypeFromJSON(json: any): V2beta1IOType {
  return V2beta1IOTypeFromJSONTyped(json, false);
}

export function V2beta1IOTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): V2beta1IOType {
  return json as V2beta1IOType;
}

export function V2beta1IOTypeToJSON(value?: V2beta1IOType | null): any {
  return value as any;
}

export function V2beta1IOTypeToJSONTyped(value: any, ignoreDiscriminator: boolean): V2beta1IOType {
  return value as V2beta1IOType;
}